Extract meta modification into module
